+++ packages/example/app.config.ts
@@ -4,7 +4,7 @@ export default ({ config }: ConfigContext): ExpoConfig => ({
   ...config,
   name: 'kstyled-example',
   slug: 'kstyled-example',
-  version: '0.4.0',
+  version: '0.4.1',
   orientation: 'portrait',
   icon: './assets/icon.png',
   userInterfaceStyle: 'automatic',